Revise flatten algorithm.
Instead of passing around a remaining_rules list that is modified and
then re-sorting the rules at the end (effectively doing a breadth-first
traversal), do a depth-first traversal that preserves ordering).

Also, hard code the renaming scheme in more places, to simplify the code,
rework how the default _should_flatten function is specified, rename
_check_ast_is_a_list_of_rules to validate_ast(), and fix a bug in
the handling of how not and paren were flattened.

At this point, everything that should be flattened is and things that
don't need to be flattened, aren't.