Make all of the compiler nodes callable.
This ensures that all of the compiler nodes can be inlined;
testing this helps ensure that I have all of the logic right.

We don't actually want all of the nodes to be inlined, because
that can make the code too hard to read; the checked-in version
contains a heuristic that inlines up to three nodes containing at
most one choice/seq/scope.