Make grammars w/ both left & right recursion work.
I'm not sure why I had both the seeds and the "blocked" rule,
but the blocked rule looked like it was implemented incorrectly,
and not sensitive to the position the rule was being called at.

I've removed it, and now things work again.

This CL also adds a few more expression grammars and test cases.